+++ commands/cloudapi-v6/datacenter.go
@@ -188,7 +188,7 @@ Required values to run command:
 	deleteCmd.AddIntFlag(constants.ArgTimeout, constants.ArgTimeoutShort, constants.DefaultTimeoutSeconds, "Timeout option for Request for Data Center deletion [seconds]")
 	deleteCmd.AddInt32Flag(cloudapiv6.ArgDepth, cloudapiv6.ArgDepthShort, cloudapiv6.DefaultDeleteDepth, cloudapiv6.ArgDepthDescription)
 
-	return core.WithConfigOverride(datacenterCmd, "compute", "")
+	return core.WithConfigOverride(datacenterCmd, []string{"cloud", "compute"}, "")
 }
 
 func PreRunDataCenterId(c *core.PreCommandConfig) error {