+++ commands/cloudapi-v6/flowlog.go
@@ -222,7 +222,7 @@ Required values to run command:
 	deleteCmd.AddIntFlag(constants.ArgTimeout, constants.ArgTimeoutShort, constants.DefaultTimeoutSeconds, "Timeout option for Request for FlowLog deletion [seconds]")
 	deleteCmd.AddInt32Flag(cloudapiv6.ArgDepth, cloudapiv6.ArgDepthShort, cloudapiv6.DefaultDeleteDepth, cloudapiv6.ArgDepthDescription)
 
-	return core.WithConfigOverride(flowLogCmd, "compute", "")
+	return core.WithConfigOverride(flowLogCmd, []string{"cloud", "compute"}, "")
 }
 
 func PreRunFlowLogList(c *core.PreCommandConfig) error {