+++ commands/cloudapi-v6/volume.go
@@ -272,7 +272,7 @@ Required values to run command:
 	deleteCmd.AddIntFlag(constants.ArgTimeout, constants.ArgTimeoutShort, constants.DefaultTimeoutSeconds, "Timeout option for Request for Volume deletion [seconds]")
 	deleteCmd.AddInt32Flag(cloudapiv6.ArgDepth, cloudapiv6.ArgDepthShort, cloudapiv6.DefaultDeleteDepth, cloudapiv6.ArgDepthDescription)
 
-	return core.WithConfigOverride(volumeCmd, "compute", "")
+	return core.WithConfigOverride(volumeCmd, []string{"cloud", "compute"}, "")
 }
 
 func PreRunVolumeList(c *core.PreCommandConfig) error {
@@ -1042,7 +1042,7 @@ Required values to run command:
 	detachVolume.AddBoolFlag(cloudapiv6.ArgAll, cloudapiv6.ArgAllShort, false, "Detach all Volumes.")
 	detachVolume.AddInt32Flag(cloudapiv6.ArgDepth, cloudapiv6.ArgDepthShort, cloudapiv6.DefaultMiscDepth, cloudapiv6.ArgDepthDescription)
 
-	return core.WithConfigOverride(serverVolumeCmd, "compute", "")
+	return core.WithConfigOverride(serverVolumeCmd, []string{"cloud", "compute"}, "")
 }
 
 func PreRunServerVolumeList(c *core.PreCommandConfig) error {